# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/oauth-clients": {
            "get": {
                "tags": [
                    "OAuth Clients"
                ],
                "summary": "Display a listing of OAuth clients.",
                "description": "Daftar aplikasi client yang terdaftar",
                "operationId": "oauthClientsIndex",
                "responses": {
                    "200": {
                        "description": "Daftar client"
                    }
                }
            },
            "post": {
                "tags": [
                    "OAuth Clients"
                ],
                "summary": "Store a newly created OAuth client.",
                "description": "Mendaftarkan aplikasi client baru",
                "operationId": "oauthClientsStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "client_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "client_secret": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "allowed_redirect_uris": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/oauth-clients/{id}": {
            "get": {
                "tags": [
                    "OAuth Clients"
                ],
                "summary": "Display the specified OAuth client.",
                "description": "Detail aplikasi client",
                "operationId": "oauthClientsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail client"
                    },
                    "404": {
                        "description": "Client tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "OAuth Clients"
                ],
                "summary": "Update the specified OAuth client.",
                "description": "Memperbarui aplikasi client",
                "operationId": "oauthClientsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "client_secret": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "allowed_redirect_uris": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client berhasil diperbarui"
                    },
                    "404": {
                        "description": "Client tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "OAuth Clients"
                ],
                "summary": "Remove the specified OAuth client.",
                "description": "Menghapus aplikasi client",
                "operationId": "oauthClientsDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Client berhasil dihapus"
                    },
                    "404": {
                        "description": "Client tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}